+++ fbw-common/src/wasm/systems/systems/src/failures/mod.rs
@@ -31,6 +31,7 @@ pub enum FailureType {
     BrakeHydraulicLeak(HydraulicColor),
     BrakeAccumulatorGasLeak,
     RadioAltimeter(usize),
+    ApuStartMotor,
 }
 
 pub struct Failure {